% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wordtree}
\alias{wordtree}
\title{wordtree}
\usage{
wordtree(text, targetWord, direction, Number_words, fileName)
}
\arguments{
\item{text}{A vector of phrases or sentances that should be used for the wordnet. You can feed the raw inputs in, we
will not use the entire sentance but will extract what we want using the `Number_words` parameter. I DO throw out all
punctuation, numbers and characters since what we're interested in is the words.}

\item{targetWord}{This is the main word which should be the "root" of the tree}

\item{direction}{This specifies in what direction the tree should "grow". To have the root word on the left, and the
tree on the right, type "suffix". In a "prefix" word tree, the root is on the right, and in a "double" word tree,
it's in the center.}

\item{Number_words}{This specifies how many words you would like the "tree" to consist of, not including the root. IE,
if we feed the wordtree two sentences, "I like the man known as Ike", and "Hey, what's up Ike?" and select `Number_words`
=2, then the sentances will be shortened to "known as Ike" and "what's up Ike"}

\item{fileName}{If you would like the wordtree to be saved somewhere, then specify the filename here.}
}
\value{
returns an html file containing the wordtree.
}
\description{
This is a simple wrapper for google's awesome Wordtree function
https://developers.google.com/chart/interactive/docs/gallery/wordtree. NOTE: You _do_ need to be online to use this
package, although apparently no data is sent to servers, as per
https://developers.google.com/chart/interactive/docs/gallery/wordtree#data-policy.
}
\examples{
sample <- c("cats are better than dogs",
"cats eat kibble",
"cats are better than hamsters",
"cats are awesome",
"cats are people too",
"cats eat mice",
"cats meowing",
"cats in the cradle",
"cats eat mice",
"cats in the cradle lyrics",
"cats eat kibble",
"cats for adoption",
"cats are family",
"cats eat mice",
"cats are better than kittens",
"cats are evil",
"cats are weird",
"cats eat mice")

wordtree(text=sample,targetWord = "cats",direction="suffix",Number_words = 4,fileName="thingie.html")
browseURL("thingie.html")

wordtree(text=sample,targetWord = "are",direction="double",Number_words=3,fileName="thingie.html")
browseURL("thingie.html")
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_extract}}
}
